refactor: clean, modernize, simplify code. remove unused code.

+++ JasperMobile/Sources/ViewControllers/JMOnboardIntroViewController/JMOnboardIntroViewController.m
@@ -143,8 +143,8 @@ static NSString * const kPageIdentifierSeemlessIntegration = @"kPageIdentifierSe
     self.homeScreenImageWidthConstraint.constant = self.homeScreenImage.image.size.width;
 
     // server screen
-    self.serverScreenImageHeightConstraint.constant = self.serverScreenImage.image.size.height * 0.2;
-    self.serverScreenImageWidthConstraint.constant = self.serverScreenImage.image.size.width * 0.2;
+    self.serverScreenImageHeightConstraint.constant = (CGFloat) (self.serverScreenImage.image.size.height * 0.2);
+    self.serverScreenImageWidthConstraint.constant = (CGFloat) (self.serverScreenImage.image.size.width * 0.2);
 }
 
 
@@ -301,8 +301,8 @@ static NSString * const kPageIdentifierSeemlessIntegration = @"kPageIdentifierSe
             CGFloat reportScreenIpadWidth = CGRectGetWidth(self.reportScreenIpadImage.frame);
             CGFloat reportScreenIphoneWidth = CGRectGetWidth(self.reportScreenIphoneImage.frame);
 
-            self.reportScreenIpadCenterXConstraint.constant = (contentViewWidth/2.0 + reportScreenIpadWidth/2.0) ;
-            self.reportScreenIphoneCenterXConstraint.constant = -(contentViewWidth/2.0 + reportScreenIphoneWidth/2.0);
+            self.reportScreenIpadCenterXConstraint.constant = (CGFloat) (contentViewWidth/2.0 + reportScreenIpadWidth/2.0);
+            self.reportScreenIphoneCenterXConstraint.constant = (CGFloat) -(contentViewWidth/2.0 + reportScreenIphoneWidth/2.0);
 
             // home screen
             self.homeScreenImageTopConstraint.constant = homeScreenImageTopConstantStartValue;
@@ -314,16 +314,16 @@ static NSString * const kPageIdentifierSeemlessIntegration = @"kPageIdentifierSe
         case JMOnboardIntroPageInstanceAccess: {
             // home screen
             self.homeScreenImageTopConstraint.constant = CGRectGetHeight(self.contentView.frame);
-            self.homeScreenImageHeightConstraint.constant = self.homeScreenImage.image.size.height * 0.2;
-            self.homeScreenImageWidthConstraint.constant = self.homeScreenImage.image.size.width * 0.2;
+            self.homeScreenImageHeightConstraint.constant = (CGFloat) (self.homeScreenImage.image.size.height * 0.2);
+            self.homeScreenImageWidthConstraint.constant = (CGFloat) (self.homeScreenImage.image.size.width * 0.2);
 
             // server screen
             self.serverScreenImageTopConstraint.constant = CGRectGetHeight(self.contentView.frame);
-            self.serverScreenImageHeightConstraint.constant = self.serverScreenImage.image.size.height * 0.2;
-            self.serverScreenImageWidthConstraint.constant = self.serverScreenImage.image.size.width * 0.2;
+            self.serverScreenImageHeightConstraint.constant = (CGFloat) (self.serverScreenImage.image.size.height * 0.2);
+            self.serverScreenImageWidthConstraint.constant = (CGFloat) (self.serverScreenImage.image.size.width * 0.2);
 
             // reports screens
-            self.reportScreenIpadCenterXConstraint.constant = self.reportScreenIpadImage.image.size.width / 6.0;
+            self.reportScreenIpadCenterXConstraint.constant = (CGFloat) (self.reportScreenIpadImage.image.size.width / 6.0);
             self.reportScreenIphoneCenterXConstraint.constant = -self.reportScreenIphoneImage.image.size.width;
 
             break;
@@ -335,8 +335,8 @@ static NSString * const kPageIdentifierSeemlessIntegration = @"kPageIdentifierSe
             CGFloat reportScreenIpadWidth = CGRectGetWidth(self.reportScreenIpadImage.frame);
             CGFloat reportScreenIphoneWidth = CGRectGetWidth(self.reportScreenIphoneImage.frame);
 
-            self.reportScreenIpadCenterXConstraint.constant = (contentViewWidth/2.0 + reportScreenIpadWidth/2.0) ;
-            self.reportScreenIphoneCenterXConstraint.constant = -(contentViewWidth/2.0 + reportScreenIphoneWidth/2.0);
+            self.reportScreenIpadCenterXConstraint.constant = (CGFloat) (contentViewWidth/2.0 + reportScreenIpadWidth/2.0);
+            self.reportScreenIphoneCenterXConstraint.constant = (CGFloat) -(contentViewWidth/2.0 + reportScreenIphoneWidth/2.0);
 
             // server screen
             self.serverScreenImageTopConstraint.constant = serverScreenImageTopConstantStartValue;
